Switched from ArrayList to Stack for history and cleaned up
Stack just makes more sense to use, as a part of that change anywhere history.size() was called was replaced by the appropriate Stack methods. I also moved out the addition of the fxmlStringAndController in isInitialAdd to the top level function and changed add method name to addIfNonDuplicate to represent the methods actual function.